tests/payment-tests.html

<!DOCTYPE html>
<html>
<head>
  <meta charset='utf-8'/>
  <title>Web Payments Tests</title>
  <link rel='stylesheet' href='../resources/testharness.css' media='all'/>
  <script src="../lib/paymentrequest.js"></script>
</head>
<body>
    <div><a href="../index.html">Index</a></div>
    <h1>Description</h1>
    <p>
      This test checks that the <code>PaymentRequest</code> runs correctly.
    </p>
    <div id='log'></div>
    <script src='../resources/testharness.js'></script>
    <script src='../resources/testharnessreport.js'></script>
    <script>
      window.__walletUrl = "payment-tests-wallet.html";

      test(function () {
          assert_true(undefined !== PaymentRequest, "PaymentRequest exists");
      }, "PaymentRequest is defined");

      /////////////////////
      // Constructor tests
      /////////////////////
      test(function () {
          assert_throws(new SyntaxError(), function() {
            new PaymentRequest();
          }, "SyntaxError expected because supportMethods is missing");
      }, "PaymentRequest constructor validates that supportedMethods argument is not optional");

      test(function () {
          assert_throws(new SyntaxError(), function() {
            new PaymentRequest("string");
          }, "SyntaxError expected because supportedMethods should be an array not a string");
      }, "PaymentRequest constructor validates that supportedMethods is an array");

      test(function () {
          assert_throws(new SyntaxError(), function() {
            new PaymentRequest([]);
          }, "SyntaxError expected because supportedMethods cannot be an empty array");
      }, "PaymentRequest constructor validates that supportedMethods is not an empty array");

      test(function () {
          assert_throws(new SyntaxError(), function() {
            new PaymentRequest(['method']);
          }, "SyntaxError expected because PaymentDetails argument is missing");
      }, "PaymentRequest constructor validates that PaymentDetails argument is not optional");

      test(function () {
          assert_throws({name:"InvalidAccessError"}, function() {
            new PaymentRequest(['method'], { });
          }, "InvalidAccessError expected because PaymentDetails.items argument is missing");
      }, "PaymentRequest constructor validates that PaymentDetails.items data is not optional");

      test(function () {
          assert_throws({name:"InvalidAccessError"}, function() {
            new PaymentRequest(['method'], { items: [] });
          }, "InvalidAccessError expected because PaymentDetails.items is of length 0");
      }, "PaymentRequest constructor validates that PaymentDetails.items must have length > 0");

      test(function () {
      new PaymentRequest(['method'], { items:[{}] });
      }, "PaymentRequest successfully constructed with valid arguments");

      /////////////////////
      // show() tests
      /////////////////////
      test(function () {
          assert_throws({name:"InvalidStateError"}, function() {
            var request = new PaymentRequest(['method'], { items:[{}] });
            request.show();
            request.show();
          }, "InvalidStateError expected when calling show() twice");
      }, "Calling show() twice throws InvalidStateError");

      (function() {
        var test = async_test("show() with no accepted payment methods rejects promise with NotSupportedError");

        var request = new PaymentRequest(
          ['notaccepted'],
          { items:[{}] }
          );

        request.show()
          .then(test.step_func(function(response) {
            assert_unreached("Promise should have been rejected with NotSupportedError but resolved instead.");
          }))
          .catch(test.step_func(function(e) {
            assert_equals(e.name,"NotSupportedError","Promise should have been rejected with NotSupportedError");
            assert_equals(request._debugOnlyState,"closed","After rejection, state should be closed");
            test.done();
        }));
      })();

      test(function () {
          assert_throws({name:"InvalidAccessError"}, function() {
          var request = new PaymentRequest(['accepted'], { }, { }, { "notaccepted":{}});
          request.show();
          }, "InvalidAccessError expected because payment method data doesn't match supported method");
      }, "show() with invalid payment data throws InvalidAccessError");

      (function() {
        var test = async_test("show() with one accepted payment methods resolves promise with that method");

        var request = new PaymentRequest(['accepted'], { items: [{}] }, { }, { "accepted":{}});

        request.show()
          .then(test.step_func(function(response) {
            assert_equals(response.methodName,"accepted","Promise should have been resolved with 'accepted' method");
            assert_equals(request._debugOnlyState,"closed","State of request should be closed");
            response.complete(true);
            assert_equals(request._debugOnlyState,"closed","State of request should be closed after complete()");
            assert_throws({name:"InvalidStateError"}, function() {
              response.complete(false);
            },"complete() should throw when if called a second time");
            test.done();
          }))
          .catch(test.step_func(function(e) {
            assert_unreached("Promise should not have been rejected. " + e.name);
          }));
      })();

      (function() {
        var test = async_test("payment request with requestShipping resolves successfully");
        var addresschange = false, optionchange = false;

        var request = new PaymentRequest(
          ['accepted'],
          { items: [{}] },
          { requestShipping:true },
          { "accepted":{}}
        );
        request.onshippingaddresschange = function(e) {
          addresschange = true;
          e.updatePaymentRequest({shippingOptions:[]});
        };
        request.onshippingoptionchange = function(e) {
          optionchange = true;
          e.updatePaymentRequest({});
        };

        request.show()
          .then(test.step_func(function(response) {
            assert_equals(response.methodName,"accepted","Promise should have been resolved with 'accepted' method");
            assert_equals(request._debugOnlyState,"closed","State of request should be accepted");
            response.complete(true);
            assert_equals(request._debugOnlyState,"closed","State of request should be closed after complete()");
            assert_throws({name:"InvalidStateError"}, function() {
              response.complete(false);
            },"complete() should throw if called a second time");
            assert_true(addresschange,"onshippingaddresschange should have been called");
            assert_true(optionchange,"onshippingoptionchange should have been called");
            test.done();
          }))
          .catch(test.step_func(function(e) {
            assert_unreached("Promise should not have been rejected. " + e.name);
          }));
      })();

      /////////////////////
      // abort() tests
      /////////////////////
      test(function () {
          assert_throws({name:"InvalidStateError"}, function() {
            var request = new PaymentRequest(['method'], { items:[{}] });
            request.abort();
          }, "InvalidStateError expected when calling abort() in created state");
      }, "Calling abort() in created state throws InvalidStateError");

    </script>
</body>
</html>
